refactor: rename and move ScalarFieldResultType
